Got things to work
The player is now able to collide with the enemy, and attack the enemy as well. However, the enemy doesn't instantly disappear and therefore the player could still run into the enemy thus killing the player too.  I think instead of instantly deleting the character we should instead place a Game Over screen.